erwaapi/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.15.1
cryptography==39.0.0
Django==5.0.11
django-cors-headers==4.6.0
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mssql-django==1.5
numpy==1.24.3
ocrlib==0.1
ollama==0.4.7
paramiko==2.12.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyodbc==5.2.0
pytz==2025.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2022.4
